Alert flow + ownership rights + exceptions handling

+++ CurrencyAlerts/src/main/java/com/example/currencyalerts/Services/CurrencyService.java
@@ -1,10 +1,16 @@
 package com.example.currencyalerts.Services;
 
 import com.example.currencyalerts.Models.Currency;
+import com.example.currencyalerts.Models.User;
 import com.example.currencyalerts.Repositories.CurrencyRepository;
+import com.example.currencyalerts.Repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
 import java.util.*;
+
+import static com.example.currencyalerts.Models.User.Role.ADMIN;
 
 @Service
 public class CurrencyService {
@@ -12,26 +18,54 @@ public class CurrencyService {
     @Autowired
     private CurrencyRepository repository;
 
+    @Autowired
+    private UserRepository userRepository;
+
+    private Set<String> unsupportedSet = new HashSet<>(Arrays.asList("ETH", "LTC", "ZKN", "MRD", "LPR", "GBZ"));
     public List<Currency> findAll() {
         return repository.findAll();
     }
 
-    public Currency addCurrency( Currency user){
-        return repository.save(user);
+    public void addCurrency(String symbol, String name, double currentPrice, int userId) throws UserUnauthorizedException, UnsupportedCurrencyCreationException {
+        Optional<User> user = userRepository.findById(userId);
+        if (user.isPresent() && user.get().getRole() == ADMIN && !unsupportedSet.contains(symbol)) {
+            Currency currency = new Currency();
+            currency.setName(name);
+            currency.setEnabled(true);
+            currency.setSymbol(symbol);
+            currency.setCurrentPrice(currentPrice);
+            currency.setCreatedTime(LocalDateTime.now());
+            repository.save(currency);
+        } else if (user.isPresent() && user.get().getRole() != ADMIN) {
+            throw new UserUnauthorizedException();
+        } else if (unsupportedSet.contains(symbol)) {
+            throw new UnsupportedCurrencyCreationException();
+        }
     }
 
-    public void deleteCurrencyById(int id){
-        repository.deleteById(id);
+    public void deleteCurrencyById(int id, int userId) throws UserUnauthorizedException {
+        Optional<User> user = userRepository.findById(userId);
+        if (user.isPresent() && user.get().getRole() == ADMIN) {
+            repository.deleteById(id);
+        } else {
+            throw new UserUnauthorizedException();
+        }
+
     }
 
     /* TODO: debug the case when an exception is thrown */
-    public Currency updateCurrency(int id, Currency currency) throws CurrencyNotFoundException{
-        Currency retrievedCurrency =  repository.findById(id).get();
-        if (retrievedCurrency != null) {
-            retrievedCurrency.setCurrentPrice(currency.getCurrentPrice());
-            return repository.save(retrievedCurrency);
-        } else {
-            throw new CurrencyNotFoundException();
+    public void updateCurrency(int id, double currentPrice, int userId) throws CurrencyNotFoundException, UserUnauthorizedException {
+        Optional<User> user = userRepository.findById(userId);
+        if (user.isPresent() && user.get().getRole() == ADMIN) {
+            Currency retrievedCurrency = repository.findById(id).get();
+            if (retrievedCurrency != null) {
+                retrievedCurrency.setCurrentPrice(currentPrice);
+                repository.save(retrievedCurrency);
+            } else {
+                throw new CurrencyNotFoundException();
+            }
+        } else if (user.isPresent() && user.get().getRole() != ADMIN) {
+            throw new UserUnauthorizedException();
         }
     }
 
@@ -39,10 +73,10 @@ public class CurrencyService {
 
         Optional<Currency> oCurrency = repository.findById(id);
 
-        if (oCurrency.isEmpty()) {
+        if (!oCurrency.isPresent()) {
             throw new CurrencyNotFoundException();
         }
 
-        throw new CurrencyNotFoundException();
+        return oCurrency.get();
     }
 }
